Add debug information
This PR adds a new debug flag (-d) that if passed in makes the "gempath analyze" also call out to "bundle info on each gem.  It then adds both the "homepath" and "summary" to the gem information.  This is very useful for debugging because sometimes it's hard to know where the gem has come from, i.e., the owner.

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>